+++ extensions/volumesnapshot/client/src/main/java/io/fabric8/volumesnapshot/client/VolumeSnapshotClient.java
@@ -35,21 +35,21 @@ import io.fabric8.volumesnapshot.client.internal.VolumeSnapshotResource;
 public interface VolumeSnapshotClient extends Client {
 
   /**
-   * API entrypoint for dealing with VolumeSnapshotClass(snapshot.storage.k8s.io/v1beta1)
+   * API entrypoint for dealing with VolumeSnapshotClass(snapshot.storage.k8s.io/v1)
    *
    * @return NonNamespaceOperation for VolumeSnapshotClass class
    */
   NonNamespaceOperation<VolumeSnapshotClass, VolumeSnapshotClassList, VolumeSnapshotClassResource> volumeSnapshotClasses();
 
   /**
-   * API entrypoint for dealing with VolumeSnapshotContent(snapshot.storage.k8s.io/v1beta1)
+   * API entrypoint for dealing with VolumeSnapshotContent(snapshot.storage.k8s.io/v1)
    *
    * @return NonNamespaceOperation for VolumeSnapshotContent class
    */
   NonNamespaceOperation<VolumeSnapshotContent, VolumeSnapshotContentList, VolumeSnapshotContentResource> volumeSnapshotContents();
 
   /**
-   * API entrypoint for dealing with VolumeSnapshot(snapshot.storage.k8s.io/v1beta1)
+   * API entrypoint for dealing with VolumeSnapshot(snapshot.storage.k8s.io/v1)
    *
    * @return MixedOperation for VolumeSnapshot class
    */