+++ extensions/volumesnapshot/client/src/main/java/io/fabric8/volumesnapshot/client/internal/VolumeSnapshotClassOperationsImpl.java
@@ -37,7 +37,7 @@ public class VolumeSnapshotClassOperationsImpl extends HasMetadataOperation<Volu
   }
 
   public VolumeSnapshotClassOperationsImpl(OperationContext context) {
-    super(context.withApiGroupName("snapshot.storage.k8s.io").withApiGroupVersion("v1beta1").withPlural("volumesnapshotclasses"));
+    super(context.withApiGroupName("snapshot.storage.k8s.io").withApiGroupVersion("v1").withPlural("volumesnapshotclasses"));
     this.type = VolumeSnapshotClass.class;
     this.listType = VolumeSnapshotClassList.class;
   }