+++ extensions/volumesnapshot/generator/cmd/generate/generate.go
@@ -18,7 +18,7 @@ package main
 import (
 	"fmt"
 	"github.com/fabric8io/kubernetes-client/generator/pkg/schemagen"
-	volumesnapshot "github.com/kubernetes-csi/external-snapshotter/v2/pkg/apis/volumesnapshot/v1beta1"
+	volumesnapshot "github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"reflect"
@@ -53,7 +53,7 @@ func main() {
 	// mapping of go packages of this module to the resulting java package
 	// optional ApiGroup and ApiVersion for the go package (which is added to the generated java class)
 	packageMapping := map[string]schemagen.PackageInformation{
-		"github.com/kubernetes-csi/external-snapshotter/v2/pkg/apis/volumesnapshot/v1beta1": {JavaPackage: "io.fabric8.volumesnapshot.api.model", ApiGroup: "snapshot.storage.k8s.io", ApiVersion: "v1beta1"},
+		"github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1": {JavaPackage: "io.fabric8.volumesnapshot.api.model", ApiGroup: "snapshot.storage.k8s.io", ApiVersion: "v1"},
 	}
 
 	// converts all packages starting with <key> to a java package using an automated scheme: